requirements.txt
asgiref==3.7.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-import-export==3.2.0
django-jazzmin==2.6.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
PyYAML==6.0
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.6.2
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
